+++ packages/guide/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### TBD
+
+#### Added
+
+- Added merit exchange
+
 ### 0.11.1 2023-09-14
 
 #### Added